+++ packages/familie-visittkort/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.20](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-visittkort@2.1.19...@navikt/familie-visittkort@2.1.20) (2021-02-08)
+
+**Note:** Version bump only for package @navikt/familie-visittkort
+
+
+
+
+
+## [2.1.19](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-visittkort@2.1.18...@navikt/familie-visittkort@2.1.19) (2021-02-08)
+
+**Note:** Version bump only for package @navikt/familie-visittkort
+
+
+
+
+
 ## [2.1.18](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-visittkort@2.1.17...@navikt/familie-visittkort@2.1.18) (2020-12-14)
 
 **Note:** Version bump only for package @navikt/familie-visittkort